refactor: LocalStrage系のダイアログをフォルダ分けて整理。
